Add MistralAIChat
Add MistralAIChat to access Mistral.ai API the same way as OpenAI, since
the Mistral API has th same API definition.
This feature needs to access a MISTRAL_API_KEY env var to be authorized
on Mistral's API
This feature uses the OpenAIChat as the parent class, and just change
the OpenAIConfig object attributes to reflect what's needed for Mistral.
As the Mistral API does not allow the use of function yet, I used the
same MissingFeatureExcetion as OllamaChat.

Be carefull on the rate limiting of Mistral.ai API : max 5 requests per
second.